refactor(gpu): split gpu_traversal.rs to satisfy 500-NLOC file limit
Devin kept flagging `gpu_traversal.rs` against CONTRIBUTING.md's
"File NLOC <= 500" mandatory quality gate. The original file had
~770 production NLOC across four concerns:

  * orchestrator + public `search_layer0` entry point
  * buffer layout (~400 NLOC of one-off allocations)
  * pipeline compilation and bind-group layouts (~155 NLOC)
  * tests

Split the file along those natural seams:

  * `gpu_traversal_buffers.rs` (473 lines / 360 NLOC) — owns
    `TraversalBuffers` struct + `new` + `create_*_bind_group` helpers.
    Now hosts `MAX_CANDIDATES_PER_ITER` since it's primarily a buffer
    sizing constant.
  * `gpu_traversal_pipelines.rs` (185 lines) — owns `compile_*_pipeline`
    constructors and matching `create_*_bind_group_layout` helpers, plus
    the `storage_entry`/`uniform_entry` const-fn builders.
  * `gpu_traversal.rs` (614 lines / 410 NLOC) — keeps orchestrator,
    `GpuTraversalContext`, `search_layer0`, `encode_*_pass`, threshold
    helpers, stats, tests.

`gpu.rs` declares both new modules (private, `mod`). All cross-module
symbols use `pub(super)` so the public surface of `velesdb-core` is
unchanged.

Verified:
  * cargo clippy --workspace --all-targets --features persistence,gpu,update-check
    --exclude velesdb-python -- -D warnings -D clippy::pedantic → 0 errors
  * cargo fmt --all --check → clean
  * 78 gpu tests pass
  * gpu_traversal.rs NLOC 410 (< 500 limit)
